zap-microservice/requirements.txt
Flask==2.2.5
python-owasp-zap-v2.4==0.0.16
pandas